private-update: return early if `ListUpdate()` fails
`ListUpdate()` returns a bool to indicate whether locking succeeded or
not, but we don't look at the return value. Teach the caller to return
early if `ListUpdate()` returns `false`.

To give a concrete example, this means that a non-root user with the
default configuration will see this on `apt-get update`:

  E: Could not open lock file /var/lib/apt/lists/lock - open (13: Permission denied)
  E: Unable to lock directory /var/lib/apt/lists/

as opposed to also seeing a couple of follow-on warnings:

  W: Problem unlinking the file /var/cache/apt/pkgcache.bin - RemoveCaches (13: Permission denied)
  W: Problem unlinking the file /var/cache/apt/srcpkgcache.bin - RemoveCaches (13: Permission denied)