acquire: hint if locking fails and we're not root
With the default configuration, if a non-root user runs `apt-get
update`, we emit errors like this:

  E: Could not open lock file /var/lib/apt/lists/lock - open (13: Permission denied)
  E: Unable to lock directory /var/lib/apt/lists/

The previous commit reduced the noise a bit, but we can also be more
helpful in our reporting. If locking fails and we're not running as
root, hint that this might be the cause of our problem.

Of course, the problem could be something else entirely and trying to
diagnose all possible causes for our problem is infeasible. But there's
a fair chance this hint will cover most of the cases, and it could be
particularly helpful to less experienced users.