Allow metadata changes to be announced and auto-accepted
Tracking a release by its codename is a rather common situation, but in
its lifetime the release may change its suite e.g. from stable to
oldstable. That can be a dangerous move if you have configuration based
on the suite name, but it is "common" enough that we shouldn't block
security updates.

To allow the best of both worlds we introduce the Future-set of metadata
so a repository can announce that it will e.g. change its Suite soon and
apt can begin issuing warnings (in the form of a notice) at that point
so a user can check its configuration and might even consider upgrading
early to help us test – a Future-Release-Notes is highly recommend.

In the case of a Suite change this change will be accepted by
downgrading the error to a notice if it was properly announced in the
last Release file apt had downloaded with the assumption that the user
has acted on the annoucement.