Consider signatures by DSA keys as weak
We deprecated and phased out SHA1 as digest algorithm. While we can
not control the key size we can at least officially phase out DSA keys
with rather similar code. There shouldn't be a lot of DSA keys still in
use, but moving directly to untrusted seems unfair, so we start with
weak here.

So similar in fact that it reuses the warning message which talks about
the "digest algorithm (DSA)" now, but it seems unfair to burden
translators with a new string which is a) rarely shown and b) probably
not very understandable for a "normal" user even if it would be talking
about "pubkey algorithm (DSA)".