source: Determine best version to download based on best binary
If a source package doesn't have a binary of the same name, the
source command always fell back to the latest version because there
is no pinning for sources.

For example, when postgresql-debversion was tested against apt,
apt source postgresql-debversion was run. But as there is no
binary of that name, only a postgresql-14-debversion in `lunar`
and a postgresql-15-debversion in `lunar-proposed`, apt downloaded
the latest version instead:

    autopkgtest [09:15:40]: @@@@@@@@@@@@@@@@@@@@ apt-source postgresql-debversion
    Get:1 http://ftpmaster.internal/ubuntu lunar-proposed/universe postgresql-debversion 1.1.2-1 (dsc) [2139 B]
    Get:2 http://ftpmaster.internal/ubuntu lunar-proposed/universe postgresql-debversion 1.1.2-1 (tar) [20.2 kB]
    Get:3 http://ftpmaster.internal/ubuntu lunar-proposed/universe postgresql-debversion 1.1.2-1 (diff) [4212 B]

With this change in place, apt will see that the proposed version
only has binaries with a pin of 100, whereas the release version
has binaries with a pin of 500, and hence download the correct
version from lunar instead.

Since we had mapping from source packages to binary versions for
quite a while use that to determine a candidate version for a
source package and then try to find a source package matching
that version.

Also adjust test-bug-998444-regex-as-target-release for the case
where no target release was given. This is exactly the same case
as above, previously 4 was picked, despite, 4 having a poin of
-10. Now 2 is picked at that's the candidate for the  binary.