Fix version finding for binary version
When we did apt source foo, and foo is a binary package, we try
to lookup the candidate and use that, but we only set the version
tag when the source version is different from the package version.

So if we have packages:

    foo         Version: 1+foo      Source: foo (= 1)
    foo         Version: 2+foo      Source: foo (= 2)
    bar         Version: 1          Source: bar (= 1)
    bar         Version: 2          Source: bar (= 2)

And we pin the "1" versions higher than 2,

    apt source foo downloads version 1
    apt source bar downloads version 2

This fixes it so that apt source bar also downloads version 1.

Note that we cannot set the VerTag if we have a RelTag, as the
RelTag must take precedence. This can be enhanced so that if there
are multiple versions matching the rel tag, the candidate takes
priority but this is not implemented so far.