Provide --online and --local flags for changelog
This basically makes Acquire::Changelogs::AlwaysOnline a tri-state
option by reinterpreting its unset value as something different than
the default false value it previously had.

It also makes it so that --local (or --offline or --no-online) still
make online requests if the changelog file doesn't exist locally like
if the package isn't installed, but what are you supposed to do, right?
It is after all, rather unlikely that you ask apt for a changelog and
want it to reply with an error. You would just prefer if it would give
you the local changelog, even if it is trimmed.