Allow command line alias to be cleanly defined in code
We have a bunch of ad-hoc aliases by having them dispatch to the same
function, but then the command line parsing doesn't recognize all of
them and so they tend to accept not the option the canonical name
provides, which can be a source of confusion at the very least and
sorta defeats the point of providing aliases to begin with.

We instead of continuing in this ad hoc fashion make the dispatch
creation the central place to define aliases and have them correctly
recognized by the code internal (as well as external in e.g. hooks).