Dashes in command names galore
Some people want 'showsrc', some expect 'show-src' and right about
everyone complains that 'autoremove' and 'auto-remove' is a thing but
there is only 'build-dep' … so, as a (probably failing) attempt to make
everyone happy (for a split second) a bunch of dashes are added and
removed in new aliases for existing commands.

Lets hope it helps.