Support changelog --no-download
As setting AlwaysOnline to false is not the same as an non-existent
AlwaysOffline option and introducing such an option would confuse this
space even more given that they (both) behave more on a always *prefer*
it rather than a always only *use* it the existing --no-download option
is now supported (not just silently ignored) to tell apt that you really
want either the offline file or an error.

This is rather edge-casey as in practice this command should really be
used interactively where --online or --local will usually do the right
thing according to what the user expects (= seeing a changelog).