Detect trimmed changelogs and prefer online in that case
The --online and --local flags can be used to manually control this, but
if you leave it up to auto-detection the file is inspected looking for
comments (at the end of the file) indicating that it was trimmed.

We only check the start of these lines to avoid hard coding the exact
command and we pick 150 as maximum line length as the longest package
name on my system is apparently 75 characters long. We could choose
longer or shorter without much issue as over-length just means we
mishandle the rest of the line as a new line and it should be really
unlikely that a) lines are that long in this file and b) that such long
lines contain one of our trigger sequences – but even if, all we do is
start a download of the online file. Could be worse.

Closes: #1024457