Make output more informative when release info changed
When the release info of a repository has changed, apt prints a
message, referring the user to the apt-secure manpage. However,
this manpage does not say how the user can resolve the situation.

This commit modifies apt so that it suggests the user to verify
the changed release info and add the --allow-releaseinfo-change
option to the command line.