WIP: Add valgrind annotations to mark mmap memory uninitialized by default
We do not want to consider the entire cache file as initialized,
that is silly. Only consider bytes we actually allocated in it
(for objects and the hash tables) as (zero) initialized and require
explicit initialization for anything else in the map.

Consider the cache hash to always be initialized as it will read
uninitialized bytes.

This still causes warnings about uninitialized bytes passed to
write() when we write the map back to a file which seemingly
can only be suppressed by a suppression file.