Introduce a REPLACE list for renamed packages in 'apt'
The t64 transition had very many renames (on 64bit archs) so that it
became really annoying to compare the NEW and REMOVED lists if they
contained only renames or also more interesting other things.

This new list (only shown by 'apt', not 'apt-get') shows the two
together and drops them from the individual NEW and REMOVED lists which
hopefully helps the user decide if seeing his favorite package 'foo'
removed is okay, because it is "just" renamed to 'bar' or if they are
actually facing a lose of the features provided by foo if they accept.

Note that this doesn't show renames achieved via transition packages nor
library transitions (like ABI changes e.g. on armhf for t64) as those do
not involve removals directly avoiding the removal scare entirely and
hence being able to be applied via 'upgrade' as the replace still counts
as a remove even if not shown in that list.