Handle new/old recommends across pkg renames
A package rename of the form newname Conflicts+Replaces+Provides oldname
in which oldname is removed looked for libapt like newname is an
entirely new package and hence "old" unsatisfied important deps of
oldname were again installed with the new name.

We now return the currently installed version of oldname as the old
version of newname instead to have the unsatisfied deps properly carry
over.