Handle new/old recommends across soname bumps
Now that we have dealt with package renames is one big set remaining:
If a soname of a library is changed, its not really a rename, sort of
the opposite as the old ABI/API disappears and stuff has to adapt to
the new one, but that is mostly a concern for developers, not for users
and both probably don't like their previous uninstalled recommends to
be added now with the "new" package as in the rename cases.

Figuring out what is a soname bump and what isn't is quite a bit of more
guesswork than for the other cases through as it is not trivial to know
the packagename with the old soname – but hopefully reality is a bit
more sane than our testcase is and we can at least detect 90% of the
cases to solve most of the problem.