Show ReplacedBy info for (auto)remove pkgs in apt
With the replace list we deal with one specific case, while we deal with
more cases while dealing with important dependencies. One area we can
reuse most of this code and help the user as well is to display the
information of which package apt considers "new" and replacing an old
one while we are (advertising) removing this old package in the
autoremove and remove lists.

Note that this info is currently not shown with the column layout as it
enlarges the columns greatly and also this extra info should probably
not be colored (or perhaps not the same way the pkg is).