Salsa-CI: Run only on scheduled runs and if manually triggered
Maintainers of apt are concerned that waiting for 20+ minutes for the CI
to complete is too long of a time, and possibly wasteful of resources as
the Salsa-CI is mostly for testing packaging and only rare commits in apt
actually change the packaging.

However, not running Salsa-CI at all before an upload risks apt having
regressions detected only after the upload. Hence allow Salsa-CI to be
run on scheduled runs (apt repository maintainers can choose the interval)
and also if triggered manually.