Add apt-gencaches.service and invoke it after upgrading apt
When we upgrade apt we bump the cache format so the cache will
be considered fully outdated, so we need to regenerate it.

This solves the case for installed systems. I don't believe this
is much of an issue for chroots or docker containers as the usual
case would be to just run the next command(s) as root too, so you
only get the delay once. Whereas on a host system, you'd go possibly
run sudo apt full-upgrade -U and then later run apt show apt as your
user account.