Handle Signed-By option in case it is a filename instead of fingerprint
Previously, we simply check if the string of 2 normalized Signed-By
value are the same, but it is allowed that both or one of the
normalized Signed-By is not a valid fingerprint, it may be a filename.

so even if the content of 2 different filename are the same, we treated
it as invalid.

Signed-off-by: Aristo Chen <aristo.chen@canonical.com>