methods: Add new sqv method
This is somewhat experimental, it has not been plumbed into
the test suite yet. We are lacking various checks for unreadable
keyrings and stuff here, that maybe is fine?

We pass everything via (temporary) files rather than pipes which
isn't exactly optimal.

Preference for sqv is configured at build time, if an sqv binary
is present.