sources.list(5): Explicitly say what the precedence order is
The order of apt sources can matter, but this man page did not
previously define whether /etc/apt/sources.list was parsed before or
after the fragment files in sources.list.d. Giving a concrete example
also provides a hint that numeric prefixes are a useful way to force a
particular sort order, and clarifies that the .list or .sources extension
is less important than lexicographic order.

Signed-off-by: Simon McVittie <smcv@debian.org>