apt_preferences(5): Specify the order in which files are parsed
The order of preference stanzas matters, but this man page did not
previously define whether /etc/apt/preferences was parsed before or
after the fragment files in preferences.d. Giving a concrete example
also provides a hint that numeric prefixes are a useful way to force a
particular sort order, and clarifies that the presence or absence of
the .pref extension is less important than lexicographic order.

Signed-off-by: Simon McVittie <smcv@debian.org>