Unset Dir::Etc::trusted
The trusted.gpg file has been deprecated since February 2022
and can no longer be managed easily without apt-key, so we
should not be falling back to it by default.

We do not remove the actual code to fall back to it, such that
code that passes Dir::Etc::trusted explicitly continues to work
for the time being (with the warnings in update), as the overhead
is now rather trivial.