solver3: Avoid bounds checks for vectors
Either we iterate over a vector, or heap sort it, or we access
it by package or version IDs (when created for the correct count
of items). Due to the high number of accesses, since all state is
recorded in vectors; bounds checks make up 20% of the solver
run-time.

Disable bounds checking for vectors only by undefining the macro
before including it and then redefining.