Test if 3.0 solver performs worse if --audit is passed
If --audit is passed, and we managed to solve a dependency,
run the 3.0 solver and check if the solver produced a worse
result.

Optimally we should be dumping the request here, instead
of resolving. The dumping is quite slow, it would be faster
if it constrained its universe to the transitive closure of
packages being changed or remaining installed.