edsp: Do not skip lines (fixes Architecture reading)
We skipped the Architecture: line because when the outer
loop got to it, the inner loop started and read the next
line after it.

Get rid of the nested loop instead, and first run a loop
to find the request line, and then a second loop to parse
the request flags.

An alternative would be to use a combined loop that
supports non-compliant protocol implementations, but
that's harder to read and we don't expect there to be
any especially on the request side.