Do not depend on sqv when pkg.apt.nosqv is required
When pkg.apt.nosqv is specified, /usr/bin/sqv can still exist making the
package depending on sqv. Now pkg.apt.nosqv is explicitly considered.

Another change is making the package explicitly depending on gpgv when
the host arch is beyond the given list, because detecting the existence
of sqv to decide whether to depend on sqv is not accurate when cross
building.

Fixes: f5a24ace (debian: Plug sqv into the package build)